backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-pymongo==2.3.0
flask-jwt-extended==4.6.0
pymongo==4.6.1
werkzeug==3.0.1
openpyxl==3.1.2
python-dotenv==1.0.0
dnspython==2.4.2
certifi==2023.11.17
